# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FinDesk AI Analyst API
servers:
  - url: /api
    description: Base API path
security:
  - bearerAuth: []
  - sessionCookie: []
    csrfToken: []
...
paths:
  /auth/session:
    post:
      operationId: createSession
      tags: [auth]
      summary: Exchange a permanent human bearer for a browser session
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Browser session created; opaque session and CSRF values are set as strict secure cookies
          content:
            application/json:
              schema:
                type: object
                required: [sessionId, principal, effectiveScopes]
                properties:
                  sessionId: { type: string }
                  principal: { type: object, additionalProperties: true }
                  effectiveScopes: { type: array, items: { type: string } }
        "401": { $ref: "#/components/responses/AuthRequired" }
        "403": { $ref: "#/components/responses/AuthForbidden" }
        "503": { $ref: "#/components/responses/AuditUnavailable" }
    delete:
      operationId: deleteSession
      tags: [auth]
      summary: Revoke the current browser session
      security:
        - sessionCookie: []
          csrfToken: []
      responses:
        "204": { description: Browser session revoked }
        "401": { $ref: "#/components/responses/AuthRequired" }
        "403": { $ref: "#/components/responses/AuthForbidden" }
        "503": { $ref: "#/components/responses/AuditUnavailable" }
  /auth/whoami:
    get:
      operationId: getCurrentPrincipal
      tags: [auth]
      summary: Return the current verified principal and effective scopes
      responses:
        "200":
          description: Current principal
          content:
            application/json:
              schema:
                type: object
                required: [principal, credentialId, effectiveScopes, authMode]
                properties:
                  principal: { type: object, additionalProperties: true }
                  credentialId: { type: string }
                  effectiveScopes: { type: array, items: { type: string } }
                  authMode: { type: string, enum: [bearer, cookie] }
                  sessionId: { type: string }
        "401": { $ref: "#/components/responses/AuthRequired" }
        "403": { $ref: "#/components/responses/AuthForbidden" }
        "503": { $ref: "#/components/responses/AuditUnavailable" }
...
  /copilot/replay/session:
    get:
      operationId: getReplaySession
      tags: [copilot]
      summary: Load replay session metadata for a symbol/date
      description: >-
        Returns canonical historical-case metadata after verified brain
        authorization and exact evidence binding. Replay never executes,
        simulates, routes, or paper-trades.
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: false
          schema:
            type: string
          description: "ISO date (YYYY-MM-DD); defaults to the symbol's available session"
        - name: caseRevisionId
          in: query
          required: true
          schema:
            type: string
          description: Exact immutable brain case revision.
        - name: evidenceHash
          in: query
          required: true
          schema:
            type: string
          description: Exact evidence hash bound to the case revision.
      responses:
        <<: *protectedResponses
        "200":
          description: Replay session metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplaySession"
        "404":
          description: No replayable session for this symbol/date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Canonical historical brain unavailable or integrity check failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /copilot/replay/event:
    get:
      operationId: getReplayEvent
      tags: [copilot]
      summary: Build the deterministic copilot event at a replay step
      description: >-
        After verified brain authorization, computes one deterministic event
        from an exact canonical case revision and evidence hash. Never returns
        order intent and never reads bundled fixtures.
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: true
          schema:
            type: string
          description: "ISO date (YYYY-MM-DD) of the replay session"
        - name: step
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
          description: "0-based replay step (0 .. totalSteps-1)"
        - name: caseRevisionId
          in: query
          required: true
          schema:
            type: string
          description: Exact immutable brain case revision.
        - name: evidenceHash
          in: query
          required: true
          schema:
            type: string
          description: Exact evidence hash bound to the case revision.
      responses:
        <<: *protectedResponses
        "200":
          description: Canonical copilot event for the replay step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopilotEvent"
        "404":
          description: Invalid replay symbol/date/step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Canonical historical brain unavailable or integrity check failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /copilot/replay/explain:
    get:
      operationId: explainReplayEvent
      tags: [copilot]
      summary: Explain the replay-step event with the read-only analyst committee
      description: >-
        Runs the same multi-agent analyst committee over the deterministic
        replay-step event. The committee only explains, critiques, and
        summarizes: it never creates signals, approves trades, overrides hard
        blocks, or invents data. The event is resolved only from the exact
        authorized canonical case revision and evidence hash.
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: true
          schema:
            type: string
        - name: step
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
        - name: caseRevisionId
          in: query
          required: true
          schema:
            type: string
          description: Exact immutable brain case revision.
        - name: evidenceHash
          in: query
          required: true
          schema:
            type: string
          description: Exact evidence hash bound to the case revision.
      responses:
        <<: *idempotentResponses
        "200":
          description: Read-only analyst committee explanation for the replay step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitteeRead"
        "404":
          description: Invalid replay symbol/date/step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Canonical historical brain unavailable or integrity check failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: permanent-application-credential
    sessionCookie:
      type: apiKey
      in: cookie
      name: mie_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
    stepUpCredential:
      type: apiKey
      in: header
      name: X-MIE-Step-Up-Authorization
  responses:
    AuthRequired:
      description: No valid application credential or browser session
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    AuthForbidden:
      description: Verified principal lacks the required kind, scope, CSRF binding, or resource permission
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    AuditUnavailable:
      description: Authentication, request audit, or idempotency persistence is unavailable
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
  schemas:
...
    ReplaySession:
      type: object
      description: Fixture-backed replay session metadata. Research/practice only; never executes, simulates, or paper-trades.
      required:
        - symbol
        - date
        - availableDates
        - dataSource
        - caseRevisionId
        - evidenceHash
        - totalSteps
        - barSeconds
        - startTime
        - endTime
      properties:
        symbol:
          type: string
        date:
          type: string
          description: "ISO date (YYYY-MM-DD) of the replayable session"
        availableDates:
          type: array
          items:
            type: string
          description: "Every ISO date this symbol can be replayed for (for the date picker)"
        dataSource:
          type: string
        caseRevisionId:
          type: string
          description: Exact immutable brain case revision used for this session.
        evidenceHash:
          type: string
          description: Exact evidence hash bound to the case revision.
        totalSteps:
          type: integer
          description: "Valid replay steps are 0-based: 0 .. totalSteps-1"
        barSeconds:
          type: integer
        startTime:
          type: integer
          description: "Epoch seconds of the first bar"
        endTime:
          type: integer
          description: "Epoch seconds of the last bar"
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
...
    CopilotBar:
      type: object
      description: A single OHLCV price bar. `t` is epoch seconds at the bar open.
      required: [t, o, h, l, c, v]
      properties:
        t:
          type: number
          description: Epoch seconds at the bar open
        o:
          type: number
          description: Open
        h:
          type: number
          description: High
        l:
          type: number
          description: Low
        c:
          type: number
          description: Close
        v:
          type: number
          description: Volume
    CopilotEvent:
      type: object
      description: Canonical deterministic copilot event. Source of truth for the analyst layer; never carries order intent.
      required:
        - eventId
        - symbol
        - timestamp
        - mode
        - dataSource
        - provenanceMode
        - alertLevel
        - l5Blocked
        - snapshot
        - marketQuality
        - triggers
        - triggerStack
        - gates
        - hardBlocks
        - riskReward
        - position
        - feedQuality
        - warnings
        - bars
      properties:
        eventId:
          type: string
        symbol:
          type: string
        timestamp:
          type: string
          description: ISO 8601 event timestamp
        mode:
          type: string
          enum: [LIVE, REPLAY, RESEARCH]
          description: "LIVE | REPLAY | RESEARCH"
        dataSource:
          type: string
          description: "Origin of the underlying bars/quotes; LIVE is alpaca_live and historical truth is canonical-case-backed only."
        provenanceMode:
          type: string
          enum: [LIVE_SIP, HISTORICAL_FIXTURE]
          description: "Verified provenance class applied by the source-policy boundary."
        caseRevisionId:
          type: ["string", "null"]
          description: "Canonical immutable case revision for historical reads; absent for LIVE_SIP."
        evidenceHash:
          type: ["string", "null"]
          description: "Evidence hash bound to the canonical historical case; absent for LIVE_SIP."
        alertLevel:
          type: ["string", "null"]
          enum: ["L1", "L2", "L3", "L4", "L5", null]
          description: "Deterministic alert ladder level L1..L5, or null"
        l5Blocked:
          type: boolean
          description: "True when a hard L5 safety block is active"
        snapshot:
          $ref: "#/components/schemas/MarketSnapshot"
        marketQuality:
          $ref: "#/components/schemas/MarketQuality"
        triggers:
          type: array
          items:
            $ref: "#/components/schemas/CopilotTrigger"
        triggerStack:
          $ref: "#/components/schemas/TriggerStack"
        gates:
          $ref: "#/components/schemas/GateVerdicts"
        hardBlocks:
          type: array
          description: "Non-overridable L5 hard-block codes; empty when not blocked"
          items:
            type: string
        riskReward:
          $ref: "#/components/schemas/RiskReward"
        position:
          $ref: "#/components/schemas/PositionRead"
        feedQuality:
          $ref: "#/components/schemas/FeedQuality"
        warnings:
          type: array
          items:
            type: string
        bars:
          type: array
          description: "OHLCV bars underlying this event, oldest first; empty on data failure."
          items:
            $ref: "#/components/schemas/CopilotBar"
    AgentRead:
      type: object
      description: One specialist analyst's read. Explanatory only; never an instruction to transact.
      required:
        - agent
        - status
        - bias
        - confidence
        - headline
        - supportingFactors
        - warnings
        - riskVerdict
        - maxRecommendation
      properties:
        agent:
          type: string
          enum:
            - technical
            - pattern
            - regime
            - order_flow
            - catalyst
            - position
            - memory
            - bull_case
            - bear_case
            - risk_critic
        status:
          type: string
          enum: [OK, DEGRADED, UNAVAILABLE]
        bias:
          type: string
          enum: [BULLISH, BEARISH, NEUTRAL, MIXED, UNKNOWN]
        confidence:
          type: number
          description: "Clamped to [0,1]"
        headline:
          type: string
        supportingFactors:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        riskVerdict:
          type: ["string", "null"]
          enum: [PASS, WARN, BLOCK, null]
          description: "Risk critic verdict; null for every other agent"
        maxRecommendation:
          type: ["string", "null"]
          enum:
            - WATCH
            - WAIT
            - AVOID
            - POSSIBLE_LONG_ZONE
            - POSSIBLE_SHORT_ZONE
            - THESIS_VALID
            - THESIS_WEAKENING
            - TRAIL_STOP
            - TAKE_PARTIALS
            - EXIT_WARNING
            - THESIS_INVALIDATED
            - DO_NOT_ADD
            - null
          description: "Risk critic recommendation ceiling; null for every other agent"
    DashboardRead:
      type: object
      description: The single synthesized, dashboard-safe read. Research/helper output only.
      required:
        - oneSentenceRead
        - recommendation
        - confidence
        - whatSupports
        - whatArguesAgainst
        - whatConfirms
        - whatInvalidates
        - positionGuidance
        - riskNotes
      properties:
        oneSentenceRead:
          type: string
        recommendation:
          type: string
          enum:
            - WATCH
            - WAIT
            - AVOID
            - POSSIBLE_LONG_ZONE
            - POSSIBLE_SHORT_ZONE
            - THESIS_VALID
            - THESIS_WEAKENING
            - TRAIL_STOP
            - TAKE_PARTIALS
            - EXIT_WARNING
            - THESIS_INVALIDATED
            - DO_NOT_ADD
        confidence:
          type: number
          description: "Clamped to [0,1]"
        whatSupports:
          type: array
          items:
            type: string
        whatArguesAgainst:
          type: array
          items:
            type: string
        whatConfirms:
          type: array
          items:
            type: string
        whatInvalidates:
          type: array
          items:
            type: string
        positionGuidance:
          type: array
          items:
            type: string
        riskNotes:
          type: array
          items:
            type: string
    CommitteeRead:
      type: object
      description: Full read-only analyst committee response for one deterministic event.
      required:
        - status
        - source
        - eventId
        - symbol
        - alertLevel
        - l5Blocked
        - provider
        - degraded
        - agents
        - dashboardRead
        - warnings
        - provenanceMode
      properties:
        status:
          type: string
          enum: [OK, FALLBACK, ERROR]
        source:
          type: string
          enum: [multi_agent_committee, deterministic_fallback]
        eventId:
          type: string
        symbol:
          type: string
        alertLevel:
          type: ["string", "null"]
          enum: ["L1", "L2", "L3", "L4", "L5", null]
        l5Blocked:
          type: boolean
        provider:
          type: string
        degraded:
          type: boolean
        agents:
          type: array
          items:
            $ref: "#/components/schemas/AgentRead"
        dashboardRead:
          $ref: "#/components/schemas/DashboardRead"
        warnings:
          type: array
          items:
            type: string
        provenanceMode:
          type: string
          enum: [LIVE_SIP, HISTORICAL_FIXTURE]
          description: "Verified provenance class inherited from the explained event."
        caseRevisionId:
          type: ["string", "null"]
          description: "Canonical immutable case revision for historical reads; absent for LIVE_SIP."
        evidenceHash:
          type: ["string", "null"]
          description: "Evidence hash bound to the canonical historical case; absent for LIVE_SIP."
    MarketSnapshot:
      type: object
      required:
        - price
        - vwap
        - rvol
        - atr
        - openingRangeHigh
        - openingRangeLow
        - volumeExpansion
        - priceLocation
        - spread
        - change1d
      properties:
        price:
          type: ["number", "null"]
        vwap:
          type: ["number", "null"]
        rvol:
          type: ["number", "null"]
        atr:
          type: ["number", "null"]
        openingRangeHigh:
          type: ["number", "null"]
        openingRangeLow:
          type: ["number", "null"]
        volumeExpansion:
          type: ["boolean", "null"]
        priceLocation:
          type: ["string", "null"]
        spread:
          type: ["number", "null"]
        change1d:
          type: ["number", "null"]
    MarketQuality:
      type: object
      required: [spreadOk, quoteFresh, liquidityOk, notes]
      properties:
        spreadOk:
          type: ["boolean", "null"]
        quoteFresh:
          type: ["boolean", "null"]
        liquidityOk:
          type: ["boolean", "null"]
        notes:
          type: ["string", "null"]
    CopilotTrigger:
      type: object
      required: [name, category, detected]
      properties:
        name:
          type: string
        category:
          type: string
          enum: [primary_edge, entry_refinement]
          description: "primary_edge | entry_refinement"
        detected:
          type: boolean
        detail:
          type: ["string", "null"]
    TriggerStack:
      type: object
      required: [stackName, category, credibility, detectedTriggers]
      properties:
        stackName:
          type: string
        category:
          type: ["string", "null"]
          enum: [primary_edge, entry_refinement, null]
        credibility:
          type: number
          description: "Deterministic credibility score in [0,1]"
        detectedTriggers:
          type: array
          items:
            type: string
    GateVerdict:
      type: object
      required: [status, reason]
      properties:
        status:
          type: string
          enum: [PASS, WARN, BLOCK]
        reason:
          type: string
    GateVerdicts:
      type: object
      required:
        - data
        - staleness
        - spread
        - marketQuality
        - credibility
        - validation
      properties:
        data:
          $ref: "#/components/schemas/GateVerdict"
        staleness:
          $ref: "#/components/schemas/GateVerdict"
        spread:
          $ref: "#/components/schemas/GateVerdict"
        marketQuality:
          $ref: "#/components/schemas/GateVerdict"
        credibility:
          $ref: "#/components/schemas/GateVerdict"
        validation:
          $ref: "#/components/schemas/GateVerdict"
    RiskReward:
      type: object
      description: Research-only structural preview. Never an order, instruction, or signal to transact.
      required:
        - direction
        - entry
        - invalidation
        - target
        - ratio
        - riskPerShare
        - notes
      properties:
        direction:
          type: ["string", "null"]
          enum: [LONG, SHORT, null]
        entry:
          type: ["number", "null"]
        invalidation:
          type: ["number", "null"]
        target:
          type: ["number", "null"]
        ratio:
          type: ["number", "null"]
        riskPerShare:
          type: ["number", "null"]
        notes:
          type: string
    PositionRead:
      type: object
      description: Manual position read for research and journaling only.
      required: [status, side, unrealizedR, thesisStatus, notes]
      properties:
        status:
          type: string
          enum: [FLAT, IN_POSITION]
        side:
          type: ["string", "null"]
          enum: [LONG, SHORT, null]
        unrealizedR:
          type: ["number", "null"]
        thesisStatus:
          type: string
          enum: [VALID, WEAKENING, INVALIDATED, UNKNOWN]
        notes:
          type: string
    FeedQuality:
      type: object
      required:
        - source
        - quoteAgeSeconds
        - barAgeSeconds
        - spreadBps
        - completeness
        - isStale
        - verdict
        - notes
      properties:
        source:
          type: string
        quoteAgeSeconds:
          type: ["number", "null"]
        barAgeSeconds:
          type: ["number", "null"]
        spreadBps:
          type: ["number", "null"]
        completeness:
          type: number
        isStale:
          type: boolean
        verdict:
          type: string
          enum: [OK, DEGRADED, BLOCKED]
        notes:
          type: ["string", "null"]
...